new mutate.cpp
